DC-10905 : Changed `item(at indexPath: IndexPath)` in `RealmStorage`  (#2)
* changed `item(at indexPath: IndexPath)` in `RealmStorage` so that it would access realm directly and get an object, instead of pulling the full result set into the `items` array.

To do this, we created a `subscript` operator in the `Section` protocol. So that everything works the way it used to, the default version of that is `items[index]`, which is part of a protocol extension to `Section`. In `RealmSection` we define `subscript` to look at `results[index]` instead.

Then in the `item(at indexPath: IndexPath)` routine for `RealmSection` we can say `return section[indexPath.item]`, which avoids the `items` array and uses Realm to get only the one single object.

* fixed complaint about swift version

* swift_version madness.

* swift-version, again.

* `swift_versions` is failing when I do a `pod update` from my client. I've no idea why, but there you go. So, temporarily commenting this out.